+++ BookMall/src/com/bookmall/dao/OrderItemDao.java
@@ -8,7 +8,7 @@ import java.sql.Connection;
 public interface OrderItemDao {
     OrderItem getOrderItemById(Connection conn, int orderItemId);
     OrderItem getOrderItemByOrderId(Connection conn, String orderId);
-    BigInteger saveOrderItem(Connection conn, OrderItem orderItem);
+    Integer saveOrderItem(Connection conn, OrderItem orderItem);
     void deleteOrderItemById(Connection conn, String orderItemId);
     void deleteOrderItemByOrderId(Connection conn, String orderId);
 }
